revamping structure
